CAS #947 -- Tweaks storage and checking of NTLMSSP_SIGNATURE in order to mitigate Casting exceptions at runtime.  See #794 for original changes and rationale.

+++ cas-server-support-spnego/src/main/java/org/jasig/cas/support/spnego/authentication/principal/SpnegoCredential.java
@@ -130,7 +130,7 @@ public final class SpnegoCredential implements Credential, Serializable {
             return false;
         }
         for (int i = 0; i < NTLM_TOKEN_MAX_LENGTH; i++) {
-            if (SpnegoConstants.NTLMSSP_SIGNATURE[i] != token[i]) {
+            if (SpnegoConstants.NTLMSSP_SIGNATURE[i].byteValue() != token[i]) {
                 return false;
             }
         }
